ml_service/requirements.txt
Flask==2.3.3
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.0.1
opencv-python==4.8.1.78
scikit-learn==1.3.0
matplotlib==3.7.2
pandas==2.0.3
requests==2.31.0